+++ changelog.md
@@ -1,4 +1,6 @@
-
+## 1.2.#
+- Added logic to process piped input. When pipe is detected, output is automatically redirected to stdout
+- Intergated with primitive_logger module
 
 ## 1.1.x